flask_proj/requirements.txt
flask==1.1.2
flask-login
flask-session
flask_socketIO
itsdangerous
jinja2
markupsafe
python-engineio
python-socketio
six
werkzeug
numpy
io
PIL
re
base64
keras
cv2
imutils
scipy
requests
ximilar.client